runlatex.sh: Add WARNEXIT variable to make "LaTeX Warning:" in .log be ignored
In testing Fedora 44's TeX Live packages, a warning in .log files
caused build failures with this message:

    LaTeX Warning: You have requested release `2026/06/01' of LaTeX,
                   but only release `2025-11-01' is available.

    Package: array 2026/01/25 v2.7a Tabular extension package (FMi)
    \col@sep=\dimen149
    ### Incompatible package(s) detected. See perfbook.log for details. ###

The root cause of this is mismatching LaTeXe2 and array.sty versions
contained in Fedora 44's texlive-base and texlive-tools packages.

This was reported at Fedora bugzilla [1], without getting much
attention of its maintainer.

As a tentative measure, add a fall-through mode in runlatex.sh which
can be enabled by saying:

    make WARNEXIT=0

Link: https://bugzilla.redhat.com/show_bug.cgi?id=2447682 [1]

+++ utilities/runlatex.sh
@@ -30,6 +30,9 @@
 #          Akira Yokosawa <akiyks@gmail.com>
 
 : ${LATEX:=pdflatex}
+: ${WARNEXIT:=1}
+
+WARN_REMAIN=0
 
 if iconv --list | grep -q -i iso-8859-1
 then
@@ -68,7 +71,10 @@ excerpt_warnings () {
 		echo "----- You can see $basename-warning.log for the warnings above. -----"
 		echo "----- If you need to, see $basename.log for details. -----"
 		rm -f $basename-warning-prev.log
-		exit 1
+		WARN_REMAIN=1
+		if [ $WARNEXIT -ne 0 ] ; then
+			exit $WARNEXIT
+		fi
 	fi
 }
 
@@ -169,8 +175,14 @@ do
 	iterate_latex
 done
 excerpt_warnings
-rm -f $basename-warning.log $basename-warning-prev.log
-echo "'$basename.pdf' is ready."
+if [ $WARNEXIT -ne 0 ] ; then
+	rm -f $basename-warning.log $basename-warning-prev.log
+fi
+if [ $WARN_REMAIN -eq 0 ] ; then
+	echo "'$basename.pdf' is ready."
+else
+	echo "'$basename.pdf' is ready, with above warning(s) ignored."
+fi
 # cleveref version check (Ubuntu 18.04 LTS has buggy one
 if grep -q -F "packageversion{0.21.1}" `kpsewhich cleveref.sty`
 then
